requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiohttp-retry==2.9.1
aiosignal==1.4.0
aiosqlite==0.21.0
amqp==5.3.1
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.3.0
billiard==4.2.1
celery==5.5.3
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==5.2.3
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-nested-routers==0.94.2
drf-spectacular==0.28.0
Faker==37.4.3
flower==2.0.1
frozenlist==1.7.0
humanize==4.13.0
idna==3.10
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
kombu==5.5.4
multidict==6.6.3
mysql-connector-python==9.3.0
packaging==25.0
parameterized==0.9.0
pika==1.3.2
prometheus_client==0.22.1
prompt_toolkit==3.0.52
propcache==0.3.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.27.1
six==1.17.0
sqlparse==0.5.3
tornado==6.5.2
twilio==9.6.5
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
yarl==1.20.1
